+++ src/components/features/application/home/about-section/components/about-content.tsx
