+++ src/components/features/application/home/about-section/components/discover-button.tsx
