+++ src/components/features/application/home/about-section/components/main-image.tsx
