+++ src/components/features/application/home/companies-section/companies-section.tsx
