+++ src/components/features/application/home/gallary-section/gallery-section.tsx
