+++ src/components/features/application/home/special-gifts-section/components/slider-offers.tsx
